Move cache items around.
