+++ crates/cache/src/lib.rs
@@ -1,10 +1,12 @@
 mod engine;
 mod helpers;
+mod item;
 mod items;
 mod reports;
 mod runfiles;
 
 pub use engine::CacheEngine;
 pub use helpers::*;
+pub use item::*;
 pub use items::*;
 pub use reports::*;
